pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-login>=0.6.3",
    "google-cloud-storage>=2.14.0,<3.0.0",
    "google-genai>=1.60.0",
    "gunicorn>=23.0.0",
    "itsdangerous>=2.2.0",
    "matplotlib>=3.10.7",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "replit-object-storage>=0.1.0",
    "reportlab>=4.4.4",
    "requests>=2.32.5",
    "sift-stack-py>=0.9.6",
    "werkzeug>=3.1.3",
]
